Attempt to fix line sensing issue
There's a bug in the line sensing block wrapper that is preventing the S3 robot from differentiating the center of a line from the absence of a line.  This fix removes two lines that erroneously inverted the results when over the center of a dark line, while testing for the presence or absence of a line.  The preceding and following lines have also been merged into a single line, for readability.

I am attempting this fix from the Github web interface, because my test server is not currently running.  This fix will need to be tested in the Demo server.